+++ compiler/src/main/kotlin/com/squareup/metro/extensions/service/ContributesServiceFir.kt
@@ -11,6 +11,7 @@ import com.squareup.metro.extensions.fir.extractScopeArgument
 import com.squareup.metro.extensions.fir.extractScopeClassId
 import com.squareup.metro.extensions.fir.findAnnotation
 import com.squareup.metro.extensions.fir.hasAnnotation
+import com.squareup.metro.extensions.fir.resolveValueParameterTypeRef
 import com.squareup.metro.extensions.squareMetroExtensionsConfig
 import dev.zacsweers.metro.compiler.MetroOptions
 import dev.zacsweers.metro.compiler.api.fir.MetroFirDeclarationGenerationExtension
@@ -22,7 +23,6 @@ import org.jetbrains.kotlin.fir.FirSession
 import org.jetbrains.kotlin.fir.declarations.DirectDeclarationsAccess
 import org.jetbrains.kotlin.fir.declarations.FirFunction
 import org.jetbrains.kotlin.fir.declarations.FirResolvePhase
-import org.jetbrains.kotlin.fir.declarations.FirValueParameter
 import org.jetbrains.kotlin.fir.declarations.builder.buildNamedFunction
 import org.jetbrains.kotlin.fir.declarations.builder.buildRegularClass
 import org.jetbrains.kotlin.fir.declarations.builder.buildValueParameter
@@ -48,14 +48,8 @@ import org.jetbrains.kotlin.fir.extensions.predicateBasedProvider
 import org.jetbrains.kotlin.fir.moduleData
 import org.jetbrains.kotlin.fir.plugin.createDefaultPrivateConstructor
 import org.jetbrains.kotlin.fir.references.builder.buildResolvedNamedReference
-import org.jetbrains.kotlin.fir.resolve.ScopeSession
-import org.jetbrains.kotlin.fir.resolve.SupertypeSupplier
-import org.jetbrains.kotlin.fir.resolve.TypeResolutionConfiguration
 import org.jetbrains.kotlin.fir.resolve.defaultType
-import org.jetbrains.kotlin.fir.resolve.providers.firProvider
 import org.jetbrains.kotlin.fir.resolve.providers.symbolProvider
-import org.jetbrains.kotlin.fir.resolve.typeResolver
-import org.jetbrains.kotlin.fir.scopes.createImportingScopes
 import org.jetbrains.kotlin.fir.scopes.kotlinScopeProvider
 import org.jetbrains.kotlin.fir.symbols.FirBasedSymbol
 import org.jetbrains.kotlin.fir.symbols.SymbolInternals
@@ -69,8 +63,6 @@ import org.jetbrains.kotlin.fir.symbols.impl.FirValueParameterSymbol
 import org.jetbrains.kotlin.fir.toEffectiveVisibility
 import org.jetbrains.kotlin.fir.toFirResolvedTypeRef
 import org.jetbrains.kotlin.fir.types.ConeKotlinType
-import org.jetbrains.kotlin.fir.types.FirResolvedTypeRef
-import org.jetbrains.kotlin.fir.types.FirTypeRef
 import org.jetbrains.kotlin.fir.types.impl.ConeClassLikeTypeImpl
 import org.jetbrains.kotlin.name.CallableId
 import org.jetbrains.kotlin.name.ClassId
@@ -596,7 +588,7 @@ public class ContributesServiceFir(session: FirSession) :
           resolvePhase = FirResolvePhase.BODY_RESOLVE
           moduleData = session.moduleData
           origin = ContributesServiceGeneratorKey.origin
-          returnTypeRef = resolveParameterTypeRef(parameter, fakeOwner)
+          returnTypeRef = resolveValueParameterTypeRef(parameter, fakeOwner, session)
           this.name = parameter.name
           symbol = FirValueParameterSymbol()
           containingDeclarationSymbol = functionSymbol
@@ -737,40 +729,6 @@ public class ContributesServiceFir(session: FirSession) :
           annotation.toAnnotationClassIdSafe(session) == ClassIds.INJECT
         }
       }
-  }
-
-  private fun resolveParameterTypeRef(
-    parameter: FirValueParameter,
-    ownerSymbol: FirRegularClassSymbol,
-  ): FirTypeRef {
-    val returnTypeRef = parameter.returnTypeRef
-    if (returnTypeRef is FirResolvedTypeRef) return returnTypeRef
-
-    val file = session.firProvider.getFirClassifierContainerFileIfAny(ownerSymbol)
-    val scopes =
-      if (file != null) {
-        createImportingScopes(file, session, ScopeSession())
-      } else {
-        emptyList()
-      }
-
-    return session.typeResolver
-      .resolveType(
-        typeRef = returnTypeRef,
-        configuration =
-          TypeResolutionConfiguration(
-            scopes = scopes,
-            containingClassDeclarations = emptyList(),
-            useSiteFile = file,
-          ),
-        areBareTypesAllowed = true,
-        isOperandOfIsOperator = false,
-        resolveDeprecations = false,
-        supertypeSupplier = SupertypeSupplier.Default,
-        expandTypeAliases = false,
-      )
-      .type
-      .toFirResolvedTypeRef()
   }
 
   private fun buildSimpleAnnotation(classId: ClassId): FirAnnotation {